+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Improved compatibility with packages that use `dplyr` without importing `data.table` (#21).
 - Added `summarize`/`summarise` verbs.
+- `right_join` gained parameter `which`.
 
 # table.express 0.3.0