+++ NEWS.md
@@ -1,4 +1,4 @@
-# table.express 0.3.0.9000
+# table.express 0.3.1
 
 - Improved compatibility with packages that use `dplyr` without importing `data.table` (#21).
 - Added `summarize`/`summarise` verbs.